Problem: can't build JNI bindings on MSVC
Solution: fixed via zproject

- added new backend for building JNI DLL on Windows
  (zproject_bindings_jni_vs20xx.gsl)

- copied existing Visual Studio templates, stripped down, as we don't
  need the full framework, only release/debug for Dynamic builds.

- fixed paths to run in bindings/jni/msvc; the standard Outputs
  property file assumes builds/msvc and don't work as such.

- run javah before building, to generate native headers; can't run from
  within msbuild (cannot parse the --classpath for some reason), so run
  as external batch file

- cannot specify all classes on single line as it exceeds maximum
  command line length; do headers one by one and skip javah when we
  already have the header (it is otherwise very slow)

- fixed numerous warnings in generated JNI C code, which only show on
  Windows

- added searching for dependent libraries so that linking works
  properly

- place output objects in bindings/jni/msvc/obj

- place output library in bindings/jni/msvc/bin

+++ bindings/jni/src/main/java/org/zeromq/czmq/Zconfig.java
@@ -179,7 +179,7 @@ public class Zconfig implements AutoCloseable{
     Returns 0 if OK, -1 if there was an error (and then does not change  
     existing data).                                                      
     */
-    native static int __reload (long self);
+    native static long __reload (long self);
     public int reload () {
         self = __reload (self);
         return 0;
